refactor: 리뷰 게시판 리펙토링
