+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/movie/review/repository/ReviewLikeRepository.java
@@ -4,5 +4,4 @@ import com.NBE4_5_SukChanHoSu.BE.domain.movie.review.entity.ReviewLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewLikeRepository extends JpaRepository<ReviewLike, Long> {
-
 }
